Extract methods for converting between *FreeMapWithCost variants
