+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/matcher/package.scala
@@ -1,5 +1,6 @@
 package coop.rchain.rholang.interpreter
 
+import cats.arrow.FunctionK
 import cats.data.{OptionT, State, StateT}
 import coop.rchain.models.Par
 import coop.rchain.rholang.interpreter.accounting.CostAccount
@@ -39,6 +40,15 @@ package object matcher {
 
       def runWithCost: (CostAccount, Option[(FreeMap, A)]) =
         s.run(Map.empty).value.run(CostAccount.zero).value
+
+      def toNonDet(): NonDetFreeMapWithCost[A] = {
+        s.mapK[StreamT[State[CostAccount, ?], ?]](
+          new FunctionK[OptionT[State[CostAccount, ?], ?], StreamT[State[CostAccount, ?], ?]] {
+            override def apply[A](
+                fa: OptionT[State[CostAccount, ?], A]): StreamT[State[CostAccount, ?], A] =
+              StreamT(fa.fold(Stream.empty[A])(single => Stream(single)))
+          })
+      }
     }
 
     implicit def toOptionalFreeMapWithCostOps[A](s: OptionalFreeMapWithCost[A]) =
@@ -90,6 +100,14 @@ package object matcher {
 
       def runWithCost: (CostAccount, Stream[(FreeMap, A)]) =
         s.run(Map.empty).value.run(CostAccount.zero).value
+
+      def toDet(): OptionalFreeMapWithCost[A] =
+        s.mapK[OptionT[State[CostAccount, ?], ?]](
+          new FunctionK[StreamT[State[CostAccount, ?], ?], OptionT[State[CostAccount, ?], ?]] {
+            override def apply[A](
+                fa: StreamT[State[CostAccount, ?], A]): OptionT[State[CostAccount, ?], A] =
+              OptionT(fa.value.map(_.headOption))
+          })
     }
 
     implicit def toNonDetFreeMapWithCostOps[A](s: NonDetFreeMapWithCost[A]) =
